+++ gltf/src/net/mgsx/gltf/shaders/emissive-only.fs.glsl
@@ -117,7 +117,7 @@ void main() {
 #endif
     
     // final frag color
-#ifdef MANUAL_SRGB
+#ifdef GAMMA_CORRECTION
     out_FragColor.rgb = vec3(pow(color,vec3(1.0/2.2)));
 #else
     out_FragColor.rgb = color;